fix: move BAISider update contents to WebUISider
