+++ react/src/components/BAISider.tsx
@@ -1,11 +1,10 @@
-import { useCustomThemeConfig } from '../helper/customThemeConfig';
 import { useWebUINavigate } from '../hooks';
 import { useThemeMode } from '../hooks/useThemeMode';
 import Flex from './Flex';
 import { Grid, SiderProps, Typography, theme } from 'antd';
 import Sider from 'antd/es/layout/Sider';
 import _ from 'lodash';
-import React, { useRef } from 'react';
+import React from 'react';
 
 export interface BAISiderProps extends SiderProps {
   logoCollapsed?: React.ReactNode;
@@ -29,10 +28,6 @@ const BAISider: React.FC<BAISiderProps> = ({
   const { Text } = Typography;
   const { isDarkMode } = useThemeMode();
   const { xs } = Grid.useBreakpoint();
-  const themeConfig = useCustomThemeConfig();
-  const logoRef = useRef<HTMLImageElement>(null);
-  const logoCollapsedRef = useRef<HTMLImageElement>(null);
-  const webuiNavigate = useWebUINavigate();
 
   return (
     <>
@@ -98,39 +93,10 @@ const BAISider: React.FC<BAISiderProps> = ({
         >
           <div className="logo-img-wrap">
             <div style={{ display: otherProps.collapsed ? 'none' : 'block' }}>
-              {themeConfig?.logo ? (
-                <img
-                  className="logo-wide"
-                  ref={logoRef}
-                  src={
-                    themeConfig?.logo?.src || '/manifest/backend.ai-text.svg'
-                  }
-                  alt={themeConfig?.logo?.alt || 'Backend.AI Logo'}
-                  onClick={() =>
-                    webuiNavigate(themeConfig?.logo?.href || '/summary')
-                  }
-                />
-              ) : (
-                logo
-              )}
+              {logo}
             </div>
             <div style={{ display: otherProps.collapsed ? 'block' : 'none' }}>
-              {themeConfig?.logo ? (
-                <img
-                  className="logo-square"
-                  ref={logoCollapsedRef}
-                  src={
-                    themeConfig?.logo?.srcCollapsed ||
-                    '/manifest/backend.ai-text.svg'
-                  }
-                  alt={themeConfig?.logo?.alt || 'Backend.AI Logo'}
-                  onClick={() =>
-                    webuiNavigate(themeConfig?.logo?.href || '/summary')
-                  }
-                />
-              ) : (
-                logoCollapsed
-              )}
+              {logoCollapsed}
             </div>
           </div>
           <div className="logo-title-wrap">
@@ -142,9 +108,7 @@ const BAISider: React.FC<BAISiderProps> = ({
                 whiteSpace: 'nowrap',
               }}
             >
-              {otherProps.collapsed
-                ? themeConfig?.logo?.logoTitleCollapsed || logoTitleCollapsed
-                : themeConfig?.logo?.logoTitle || logoTitle}
+              {otherProps.collapsed ? logoTitleCollapsed : logoTitle}
             </Typography.Text>
           </div>
         </Flex>
