+++ react/src/components/MainLayout/WebUISider.tsx
@@ -1,3 +1,4 @@
+import { useCustomThemeConfig } from '../../helper/customThemeConfig';
 import { useSuspendedBackendaiClient, useWebUINavigate } from '../../hooks';
 import { useCurrentUserRole } from '../../hooks/backendai';
 import BAIMenu from '../BAIMenu';
@@ -22,7 +23,7 @@ import {
 } from '@ant-design/icons';
 import { theme, MenuProps, Typography } from 'antd';
 import _ from 'lodash';
-import React from 'react';
+import React, { useRef } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useLocation } from 'react-router-dom';
 
@@ -41,6 +42,9 @@ const WebUISider: React.FC<WebUISiderProps> = (props) => {
   const fasttrackEndpoint = baiClient?._config?.fasttrackEndpoint ?? null;
   const blockList = baiClient?._config?.blockList ?? null;
   const inactiveList = baiClient?._config?.inactiveList ?? null;
+  const themeConfig = useCustomThemeConfig();
+  const logoRef = useRef<HTMLImageElement>(null);
+  const logoCollapsedRef = useRef<HTMLImageElement>(null);
 
   const generalMenu: MenuProps['items'] = [
     {
@@ -165,25 +169,47 @@ const WebUISider: React.FC<WebUISiderProps> = (props) => {
   return (
     <BAISider
       logo={
-        <img
-          alt="Backend.AI Logo"
-          className="logo-wide"
-          src={'/manifest/backend.ai-text.svg'}
-          style={{ width: 191, height: 32, cursor: 'pointer' }}
-          onClick={() => webuiNavigate('/summary')}
-        />
+        themeConfig?.logo ? (
+          <img
+            className="logo-wide"
+            ref={logoRef}
+            src={themeConfig?.logo?.src || '/manifest/backend.ai-text.svg'}
+            alt={themeConfig?.logo?.alt || 'Backend.AI Logo'}
+            onClick={() => webuiNavigate(themeConfig?.logo?.href || '/summary')}
+          />
+        ) : (
+          <img
+            alt="Backend.AI Logo"
+            className="logo-wide"
+            src={'/manifest/backend.ai-text.svg'}
+            style={{ width: 191, height: 32, cursor: 'pointer' }}
+            onClick={() => webuiNavigate('/summary')}
+          />
+        )
       }
       logoCollapsed={
-        <img
-          alt="Backend.AI Logo"
-          className="logo-square"
-          src={'/manifest/backend.ai-brand-simple.svg'}
-          style={{ width: 48, height: 32, cursor: 'pointer' }}
-          onClick={() => webuiNavigate('/summary')}
-        />
+        themeConfig?.logo ? (
+          <img
+            className="logo-square"
+            ref={logoCollapsedRef}
+            src={
+              themeConfig?.logo?.srcCollapsed || '/manifest/backend.ai-text.svg'
+            }
+            alt={themeConfig?.logo?.alt || 'Backend.AI Logo'}
+            onClick={() => webuiNavigate(themeConfig?.logo?.href || '/summary')}
+          />
+        ) : (
+          <img
+            alt="Backend.AI Logo"
+            className="logo-square"
+            src={'/manifest/backend.ai-brand-simple.svg'}
+            style={{ width: 48, height: 32, cursor: 'pointer' }}
+            onClick={() => webuiNavigate('/summary')}
+          />
+        )
       }
-      logoTitle="WebUI"
-      logoTitleCollapsed="WebUI"
+      logoTitle={themeConfig?.logo?.logoTitle || 'WebUI'}
+      logoTitleCollapsed={themeConfig?.logo?.logoTitleCollapsed || 'WebUI'}
       bottomText={
         props.collapsed ? null : (
           <>
